refactor native mode
