+++ nicegui/native/__init__.py
@@ -1,0 +1,12 @@
+from .native import WindowProxy, method_queue, response_queue
+from .native_config import NativeConfig
+from .native_mode import activate, find_open_port
+
+__all__ = [
+    'activate',
+    'find_open_port',
+    'method_queue',
+    'NativeConfig',
+    'response_queue',
+    'WindowProxy',
+]
